refactor(config): isolate live policy impact execution
